feat!: cleanup sp entrypoint
BREAKING CHANGES: waitForPullStatus renamed to waitForPullPieces
