+++ packages/synapse-core/src/errors/index.ts
@@ -15,3 +15,4 @@ export * from './pay.ts'
 export * from './pdp.ts'
 export * from './piece.ts'
 export * from './pull.ts'
+export * from './warm-storage.ts'
