+++ packages/synapse-core/src/sp/index.ts
@@ -11,13 +11,12 @@
 
 export { AbortError, NetworkError, TimeoutError } from 'iso-web/http'
 export * from './add-pieces.ts'
-export * from './data-sets.ts'
+export * from './create-dataset.ts'
+export * from './create-dataset-add-pieces.ts'
+export * from './find-piece.ts'
 export * from './get-data-set.ts'
-export * from './pull.ts'
+export * from './ping.ts'
+export * from './pull-pieces.ts'
 export * from './schedule-piece-deletion.ts'
-export type { deletePiece, PullPieceStatus, PullResponse, PullStatus, UploadPieceStreamingData } from './sp.ts'
-export { findPiece, ping, uploadPiece, uploadPieceStreaming } from './sp.ts'
 export * from './upload.ts'
-export * from './wait-for-add-pieces.ts'
-export * from './wait-for-create-dataset.ts'
-export * from './wait-for-create-dataset-add-pieces.ts'
+export * from './upload-streaming.ts'
