+++ packages/synapse-core/src/sp/schedule-piece-deletion.ts
@@ -1,6 +1,50 @@
-import type { Account, Chain, Client, Transport } from 'viem'
+import { type AbortError, HttpError, type NetworkError, request, type TimeoutError } from 'iso-web/http'
+import type { Account, Chain, Client, Hex, Transport } from 'viem'
+import { DeletePieceError } from '../errors/pdp.ts'
 import { signSchedulePieceRemovals } from '../typed-data/sign-schedule-piece-removals.ts'
-import * as SP from './sp.ts'
+import { RETRY_CONSTANTS } from '../utils/constants.ts'
+
+export namespace deletePiece {
+  export type OptionsType = {
+    serviceURL: string
+    dataSetId: bigint
+    pieceId: bigint
+    extraData: Hex
+  }
+  export type OutputType = {
+    hash: Hex
+  }
+  export type ErrorType = DeletePieceError | TimeoutError | NetworkError | AbortError
+}
+
+/**
+ * Delete a piece from a data set on the PDP API.
+ *
+ * DELETE /pdp/data-sets/{dataSetId}/pieces/{pieceId}
+ *
+ * @param options - {@link deletePiece.OptionsType}
+ * @returns Hash of the delete operation {@link deletePiece.OutputType}
+ * @throws Errors {@link deletePiece.ErrorType}
+ */
+export async function deletePiece(options: deletePiece.OptionsType): Promise<deletePiece.OutputType> {
+  const { serviceURL, dataSetId, pieceId, extraData } = options
+  const response = await request.json.delete<{ txHash: Hex }>(
+    new URL(`pdp/data-sets/${dataSetId}/pieces/${pieceId}`, serviceURL),
+    {
+      body: { extraData },
+      timeout: RETRY_CONSTANTS.MAX_RETRY_TIME,
+    }
+  )
+
+  if (response.error) {
+    if (HttpError.is(response.error)) {
+      throw new DeletePieceError(await response.error.response.text())
+    }
+    throw response.error
+  }
+
+  return { hash: response.result.txHash }
+}
 
 export namespace schedulePieceDeletion {
   export type OptionsType = {
@@ -13,8 +57,8 @@ export namespace schedulePieceDeletion {
     /** The service URL of the PDP API. */
     serviceURL: string
   }
-  export type OutputType = SP.deletePiece.OutputType
-  export type ErrorType = SP.deletePiece.ErrorType
+  export type OutputType = deletePiece.OutputType
+  export type ErrorType = deletePiece.ErrorType
 }
 
 /**
@@ -55,7 +99,7 @@ export async function schedulePieceDeletion(
   client: Client<Transport, Chain, Account>,
   options: schedulePieceDeletion.OptionsType
 ): Promise<schedulePieceDeletion.OutputType> {
-  return SP.deletePiece({
+  return deletePiece({
     serviceURL: options.serviceURL,
     dataSetId: options.dataSetId,
     pieceId: options.pieceId,
